Pulls the intelligent keyboard into its own file.
